fix: create new model

+++ src/modules/workshop/components/Modal/UnsaveWarningModal/index.tsx
@@ -27,8 +27,8 @@ const UnsaveWarningModal = ({ show, setIsOpen }: Props) => {
         <div className='flex items-center justify-between mt-6'>
           <button
             onClick={async () => {
-              const res = await handleSaveFile()
-              if (res !== 'failed') {
+              const res = await handleSaveFile('save-exit')
+              if (!!res && res !== 'failed') {
                 createProject()
                 deleteAlls()
                 setIsOpen(false)
